chore: move functions to init.kt and utils.kt
